server/requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic
python-jose[cryptography]==3.3.0
passlib
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==42.0.0
httpx==0.26.0
google-generativeai==0.3.2
bcrypt==4.1.2